security: remove the security_netlink_recv hook as it is equivalent to capable()
Once upon a time netlink was not sync and we had to get the effective
capabilities from the skb that was being received.  Today we instead get
the capabilities from the current task.  This has rendered the entire
purpose of the hook moot as it is now functionally equivalent to the
capable() call.

Change-Id: I9104325a606383bb0eb8bf6d8d3786d64f1bbe9c
Signed-off-by: Eric Paris <eparis@redhat.com>